Lots of Refactoring!!! Most Notably Mecanum wheels (and its interface) have been split up into two different classes. This is to separate the functionality of encoders and regular driving, as to make the encoders more optional. Most of the underlying execution of the movement has also been reworked to be much more organized with about over 9000% less code duplication. Additionally, by the way it has been reworked, it is possible to use the class to a limited degree to control the motors in TeleOp, rather than only programmed for Autonomous (See "TestDriving" for example.)
Some other random refactoring was also done, such as changing the conditions in autonomous to use our new condition classes. (It's basically just as an example for now, the class will inevitably be removed at some point in our refactoring posses, probably once we get to reworking the structure of the OpModes themselves.)

A couple of files were moved around or folders renamed so that the package better represented what the classes did.
(Well hopefully I dont find any typos in the next 5 minutes...)